Refactoring the sources to organize packages
