+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/StorageUnitBlue.java
@@ -1,14 +1,15 @@
-package org.firstinspires.ftc.teamcode.Configs.newConfig;
+package org.firstinspires.ftc.teamcode.auto;
 
-import static org.firstinspires.ftc.teamcode.Configs.newConfig.NewAutoDriveUtils.encoderDrive;
-import static org.firstinspires.ftc.teamcode.Configs.newConfig.NewAutoDriveUtils.logData;
-import static org.firstinspires.ftc.teamcode.Configs.newConfig.NewAutoDriveUtils.logLine;
+import static org.firstinspires.ftc.teamcode.common.utils.DriveUtils.encoderDrive;
+import static org.firstinspires.ftc.teamcode.common.utils.DriveUtils.logLine;
 
 import android.annotation.SuppressLint;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.util.ElapsedTime;
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+
+import org.firstinspires.ftc.teamcode.config.HardwareNew;
+import org.firstinspires.ftc.teamcode.common.BaseNewOpMode;
 
 /**
  * This class defines the warehouse driving.
