+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/WarehouseDriveBlue.java
@@ -1,41 +1,45 @@
-package org.firstinspires.ftc.teamcode.Configs.oldConfig.selfDrive;
+package org.firstinspires.ftc.teamcode.auto;
 
-import static org.firstinspires.ftc.teamcode.Configs.oldConfig.selfDrive.AutoDriveUtils.encoderDrive;
-import static org.firstinspires.ftc.teamcode.Configs.oldConfig.selfDrive.AutoDriveUtils.logData;
-import static org.firstinspires.ftc.teamcode.Configs.oldConfig.selfDrive.AutoDriveUtils.logLine;
+import static org.firstinspires.ftc.teamcode.common.utils.DriveUtils.encoderDrive;
+import static org.firstinspires.ftc.teamcode.common.utils.DriveUtils.logData;
+import static org.firstinspires.ftc.teamcode.common.utils.DriveUtils.logLine;
+
 
 import android.annotation.SuppressLint;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-import org.firstinspires.ftc.teamcode.Configs.oldConfig.BaseOpMode;
-import org.firstinspires.ftc.teamcode.Configs.oldConfig.Hardware2;
+import org.firstinspires.ftc.teamcode.config.HardwareNew;
+import org.firstinspires.ftc.teamcode.common.BaseNewOpMode;
+
 
 /**
  * This class defines the warehouse driving.
  * @author aryansinha
+ * kind-of in a less prominent way
+ * @soon-to-be-author karthikperi
  */
-@Autonomous(name="Warehouse Driving")
-public class WarehouseDrive extends BaseOpMode {
-    private final Hardware2 robot;
+@Autonomous(name="Warehouse Drive Blue")
+public class WarehouseDriveBlue extends BaseNewOpMode {
+    private final HardwareNew robot;
     private final ElapsedTime runtime = new ElapsedTime();
 
-    public WarehouseDrive() {
+    public WarehouseDriveBlue() {
         // Start with the encoders.
-        robot = new Hardware2(true);
+        robot = new HardwareNew(true);
     }
 
     /**
      * {@inheritDoc}
      */
-    public Hardware2 getRobot() {
+    public HardwareNew getRobot() {
         return robot;
     }
 
     /**
      * {@inheritDoc}
-     */
+     * */
     @SuppressLint("DefaultLocale")
     @Override
     public void runOpMode() throws InterruptedException {
@@ -49,28 +53,28 @@ public class WarehouseDrive extends BaseOpMode {
         //reset Runtime
         runtime.reset();
 
-        encoderDrive(this,1,(double)49,(double)49, 5);
+        encoderDrive(this,1,(double)26,(double)26, 5);
+        robot.getArm().setPower(0.5);
+        sleep(5000);
+        robot.getArm().setPower(0);
+        robot.getLeftClaw().setPosition(0.5);
+        robot.getRightClaw().setPosition(0.5);
+        encoderDrive(this, 1, (double)-20, (double)-20, 5);
         robot.turnLeft(90, 0.2);
-        encoderDrive(this, 1, 2, 2, 1);
-        robot.getClawServo().setPosition(0.1);
         encoderDrive(this,-1,(double)49,(double)49, 5);
         robot.turnRight(90, 0.2);
         encoderDrive(this, -1, 2, 2, 1);
-        robot.getClawServo().setPosition(0.1);
+
         encoderDrive(this, 1.0, 25,25,5);
         logData(this, "Encoder Counts", String.format("Current Position is %7d :%7d",
                 robot.getLeftDrive().getCurrentPosition(),
                 robot.getRightDrive().getCurrentPosition()));
 
 
+
         robot.turnLeft(90, 0.1);
         encoderDrive(this, 1, 80, 80, 5);
-        robot.getClaw().setPower(1);
         sleep(500);
-        robot.getClaw().setPower(0);
-        robot.getClawServo().setPosition(1);
-        robot.getClawServo().setPosition(0);
-
 
     }
 }
