+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/BaseNewOpMode.java
@@ -1,19 +1,18 @@
-package org.firstinspires.ftc.teamcode.Configs.oldConfig;
+package org.firstinspires.ftc.teamcode.common;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+
+import org.firstinspires.ftc.teamcode.config.HardwareNew;
 
 /**
  * This class is the base for handling all robot movements.
  * @author aryansinha
  */
-public abstract class BaseOpMode extends LinearOpMode {
+public abstract class BaseNewOpMode extends LinearOpMode {
     /**
      * Returns the underlying robot.
      *
      * @return The underlying robot.
      */
-    public Hardware2 getRobot() {
-        return null;
-    }
-
+    public abstract HardwareNew getRobot();
 }
