+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/obsoleted/gen1/BaseOpMode.java
@@ -1,20 +1,18 @@
-package org.firstinspires.ftc.teamcode.Configs.newConfig;
+package org.firstinspires.ftc.teamcode.obsoleted.gen1;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-
-import org.firstinspires.ftc.teamcode.Configs.newConfig.HardwareNew;
 
 /**
  * This class is the base for handling all robot movements.
  * @author aryansinha
  */
-public abstract class BaseNewOpMode extends LinearOpMode {
+public abstract class BaseOpMode extends LinearOpMode {
     /**
      * Returns the underlying robot.
      *
      * @return The underlying robot.
      */
-    public HardwareNew getRobot() {
+    public Hardware2 getRobot() {
         return null;
     }
